added logic to exclude folders

+++ src/PhotoSyncManager/Models/LibraryProcessor.cs
@@ -14,7 +14,7 @@ namespace PhotoSyncManager.Models
     {
         public IEnumerable<PhotoRecord> Run(PhotoLibrary library)
         {
-            var files = new GetPhotoFilesQuery().Run(library.SourceFolder);
+            var files = new GetPhotoFilesQuery().Run(library);
             var sourcePathLength = library.SourceFolder.Length;
             var records = new ConcurrentBag<PhotoRecord>();
             var exceptions = new ConcurrentBag<Exception>();
@@ -25,29 +25,33 @@ namespace PhotoSyncManager.Models
                 {
                     var relativePath = file.FullName.Remove(0, sourcePathLength).TrimStart(new[] { '\\' });
                     using var context = PhotoSyncContextFactory.Make(library.DestinationFullPath);
-                    if (context.Photos.Any(x => x.RelativePath == relativePath))
+                    var excludedFolders = context.ExcludeFolders.Select(x => x.RelativePath);
+                    if (!this.IsInExcludedFolder(excludedFolders, relativePath))
                     {
-                        var photo = context.Photos.FirstOrDefault(x => x.RelativePath == relativePath);
-                        if (photo != null)
+                        if (context.Photos.Any(x => x.RelativePath == relativePath))
                         {
-                            var sourceFilePath = Path.Combine(library.SourceFolder, relativePath);
-                            if (!File.Exists(sourceFilePath))
+                            var photo = context.Photos.FirstOrDefault(x => x.RelativePath == relativePath);
+                            if (photo != null)
                             {
-                                context.Photos.Remove(photo);
-                                context.SaveChanges();
-                            }
-                            else
-                            {
-                                records.Add(this.MakeRecord(photo, file));
+                                var sourceFilePath = Path.Combine(library.SourceFolder, relativePath);
+                                if (!File.Exists(sourceFilePath))
+                                {
+                                    context.Photos.Remove(photo);
+                                    context.SaveChanges();
+                                }
+                                else
+                                {
+                                    records.Add(this.MakeRecord(photo, file));
+                                }
                             }
                         }
-                    }
-                    else
-                    {
-                        var photo = new Photo { RelativePath = relativePath };
-                        context.Photos.Add(photo);
-                        context.SaveChanges();
-                        records.Add(this.MakeRecord(photo, file));
+                        else
+                        {
+                            var photo = new Photo { RelativePath = relativePath };
+                            context.Photos.Add(photo);
+                            context.SaveChanges();
+                            records.Add(this.MakeRecord(photo, file));
+                        }
                     }
                 }
                 catch (Exception ex)
@@ -59,6 +63,19 @@ namespace PhotoSyncManager.Models
             return exceptions.IsEmpty
                 ? records
                 : throw new AggregateException(exceptions);
+        }
+
+        private bool IsInExcludedFolder(IEnumerable<string> excludedFolders, string relativePath)
+        {
+            foreach (var folder in excludedFolders)
+            {
+                if (relativePath.StartsWith(folder))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         private PhotoRecord MakeRecord(Photo photo, FileInfo file)
